requirements.txt
flask==3.0.2
flask-cors==4.0.0
mysql-connector-python==8.3.0
openai==1.12.0
Pillow==10.2.0
numpy==1.24.3
torch==2.1.0
transformers==4.35.2
google-generativeai==0.3.2
requests==2.31.0
ftfy==6.1.3
regex==2023.10.3
tqdm==4.66.1
clip @ git+https://github.com/openai/CLIP.git 